DataGrid: move internal component so subfolder
