+++ Source/Extensions/Blazorise.DataGrid/Internal/_DataGridAggregateRow.razor.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Microsoft.AspNetCore.Components;
 #endregion
 
-namespace Blazorise.DataGrid;
+namespace Blazorise.DataGrid.Internal;
 
 public partial class _DataGridAggregateRow<TItem> : BaseDataGridComponent
 {
