+++ Source/Extensions/Blazorise.DataGrid/Internal/_DataGridCancelBatchCommand.razor
