+++ Source/Extensions/Blazorise.DataGrid/Internal/_DataGridCancelBatchCommand.razor.cs
@@ -4,7 +4,7 @@ using Blazorise.Localization;
 using Microsoft.AspNetCore.Components;
 #endregion
 
-namespace Blazorise.DataGrid;
+namespace Blazorise.DataGrid.Internal;
 
 public partial class _DataGridCancelBatchCommand<TItem> : ComponentBase, IDisposable
 {
