+++ Source/Extensions/Blazorise.DataGrid/Internal/_DataGridDetailRow.razor.cs
@@ -4,20 +4,17 @@ using System.Linq;
 using Microsoft.AspNetCore.Components;
 #endregion
 
-namespace Blazorise.DataGrid;
+namespace Blazorise.DataGrid.Internal;
 
-public abstract class _BaseDataGridFullColumnSpanRow<TItem> : BaseDataGridComponent
+public abstract class _BaseDataGridDetailRow<TItem> : BaseDataGridComponent
 {
     #region Properties
-
-    protected override bool ShouldRender()
-        => RenderUpdates;
 
     protected bool HasCommandColumn
         => Columns.Any( x => x.ColumnType == DataGridColumnType.Command );
 
     protected int ColumnSpan
-        => Columns.Count - ( HasCommandColumn && !ParentDataGrid.Editable ? 1 : 0 );
+        => Columns.Where( x => x.Displayable ).Count() - ( HasCommandColumn && !ParentDataGrid.Editable ? 1 : 0 );
 
     /// <summary>
     /// Item associated with the data set.
@@ -35,8 +32,6 @@ public abstract class _BaseDataGridFullColumnSpanRow<TItem> : BaseDataGridCompon
     [CascadingParameter] public DataGrid<TItem> ParentDataGrid { get; set; }
 
     [Parameter] public RenderFragment ChildContent { get; set; }
-
-    [Parameter] public bool RenderUpdates { get; set; }
 
     #endregion
 }
