+++ Source/Extensions/Blazorise.DataGrid/Internal/_DataGridModal.razor.cs
@@ -7,7 +7,7 @@ using Blazorise.Localization;
 using Microsoft.AspNetCore.Components;
 #endregion
 
-namespace Blazorise.DataGrid;
+namespace Blazorise.DataGrid.Internal;
 
 public partial class _DataGridModal<TItem> : BaseAfterRenderComponent, IDisposable
 {
