+++ Source/Extensions/Blazorise.DataGrid/Internal/_DataGridNewCommand.razor.cs
@@ -4,9 +4,9 @@ using Blazorise.Localization;
 using Microsoft.AspNetCore.Components;
 #endregion
 
-namespace Blazorise.DataGrid;
+namespace Blazorise.DataGrid.Internal;
 
-public partial class _DataGridSaveBatchCommand<TItem> : ComponentBase, IDisposable
+public abstract class _BaseDataGridNewCommand<TItem> : ComponentBase, IDisposable
 {
     protected override void OnInitialized()
     {
@@ -34,5 +34,5 @@ public partial class _DataGridSaveBatchCommand<TItem> : ComponentBase, IDisposab
     /// </summary>
     [CascadingParameter] public DataGrid<TItem> ParentDataGrid { get; set; }
 
-    [Parameter] public EventCallback SaveBatch { get; set; }
+    [Parameter] public EventCallback New { get; set; }
 }
