+++ Source/Extensions/Blazorise.DataGrid/Internal/_DataGridRowCommand.razor.cs
@@ -6,7 +6,7 @@ using Blazorise.Localization;
 using Microsoft.AspNetCore.Components;
 #endregion
 
-namespace Blazorise.DataGrid;
+namespace Blazorise.DataGrid.Internal;
 
 public abstract class _BaseDataGridRowCommand<TItem> : ComponentBase, IDisposable
 {
