+++ Source/Extensions/Blazorise.DataGrid/_Imports.razor
@@ -1,2 +1,3 @@
 ﻿@using Microsoft.AspNetCore.Components.Web
+@using Blazorise.DataGrid.Internal
 @using static Blazorise.DataGrid.Constants
